+++ djangocms_blog/cms_apps.py
@@ -11,6 +11,7 @@ from .cms_menus import BlogCategoryMenu
 from .settings import get_setting
 
 
+@apphook_pool.register
 class BlogApp(AutoCMSAppMixin, CMSConfigApp):
     name = _('Blog')
     urls = ['djangocms_blog.urls']
@@ -30,5 +31,4 @@ class BlogApp(AutoCMSAppMixin, CMSConfigApp):
     }
 
 
-apphook_pool.register(BlogApp)
 BlogApp.setup()