+++ djangocms_blog/liveblog/models.py
@@ -3,8 +3,9 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 import json
 
+import django
 from channels import Group
-from cms.models import itemgetter, now, python_2_unicode_compatible
+from cms.models import CMSPlugin, itemgetter, now, python_2_unicode_compatible
 from cms.utils.plugins import reorder_plugins
 from django.db import models
 from django.template import Context
@@ -47,7 +48,7 @@ class LiveblogInterface(models.Model):
             items.extend(
                 model.objects.filter(placeholder=self.placeholder).values('pk', 'post_date')
             )
-        order = [item['pk'] for item in sorted(items, key=itemgetter('post_date'))]
+        order = reversed([item['pk'] for item in sorted(items, key=itemgetter('post_date'))])
         reorder_plugins(self.placeholder, None, self.language, order)
 
     @property
@@ -87,6 +88,12 @@ class Liveblog(LiveblogInterface, AbstractText):
     """
     Basic liveblog plugin model
     """
+    if django.VERSION >= (1, 10):
+        cmsplugin_ptr = models.OneToOneField(
+            CMSPlugin,
+            related_name='%(app_label)s_%(class)s', primary_key=True,
+            parent_link=True, on_delete=models.CASCADE
+        )
     title = models.CharField(_('title'), max_length=255)
     image = FilerImageField(
         verbose_name=_('image'), blank=True, null=True, on_delete=models.SET_NULL,