+++ djangocms_blog/migrations/0017_thumbnail_move.py
@@ -13,9 +13,6 @@ class Migration(migrations.Migration):
             ('filer', '0003_thumbnailoption'),
             ('cmsplugin_filer_image', '0003_mv_thumbnail_option_to_filer_20160119_1720'),
         ]
-        run_before = [
-            ('cmsplugin_filer_image', '0004_auto_20160120_0950'),
-        ]
 
         operations = [
             migrations.AlterField(